Replace struct Coordinate on Vector2. Fix bug with async spawn enemy. Fix bug with stop of moving GameCharacter

+++ TowerDefense.cs
@@ -94,7 +94,7 @@ namespace TowerDefense
             playerController.ListenEvents();
 
             //Data
-            var dt = gameTime.ElapsedGameTime.TotalMilliseconds / 1000;
+            var dt = (float)gameTime.ElapsedGameTime.TotalMilliseconds / 1000;
             sessionData.Update(dt);
 
             base.Update(gameTime);
